# shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/google": {
      "post": {
        "operationId": "AuthController_google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/admin-auth/login": {
      "post": {
        "operationId": "AdminAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/admin-auth/totp": {
      "post": {
        "operationId": "AdminAuthController_verifyTotp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/admin-auth/totp-setup": {
      "post": {
        "operationId": "AdminAuthController_setupTotp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-auth"
        ]
      }
    },
...
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsersController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/ledger": {
      "get": {
        "operationId": "AdminUsersController_ledger",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/adjust-balance": {
      "post": {
        "operationId": "AdminUsersController_adjustBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustBalanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/ban": {
      "post": {
        "operationId": "AdminUsersController_ban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/unban": {
      "post": {
        "operationId": "AdminUsersController_unban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/admins": {
      "get": {
        "operationId": "AdminAdminsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminAdminsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/admins/{id}/disable": {
      "post": {
        "operationId": "AdminAdminsController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/fraud-flags/{id}/resolve": {
      "post": {
        "operationId": "AdminFraudController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveFraudDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/game/round-start": {
      "post": {
        "operationId": "GameController_roundStart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoundStartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "game"
        ]
      }
    },
    "/api/game/round-complete": {
      "post": {
        "operationId": "GameController_roundComplete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoundCompleteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "game"
        ]
      }
    },
...
    "/api/bonus/{type}": {
      "get": {
        "operationId": "BonusController_state",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bonus"
        ]
      }
    },
    "/api/bonus/{type}/play": {
      "post": {
        "operationId": "BonusController_play",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "bonus"
        ]
      }
    }
  },
  "info": {
    "title": "Cash Raja API",
    "description": "Coin-rewards backend: auth, wallet/ledger, offers & ads postbacks, engagement (game/streak/scratch-spin/referral), redemptions, notifications, fraud, and the admin API.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "default"
    }
  ],
  "components": {
    "securitySchemes": {
      "user": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "App-user access token (aud=app)"
      },
      "admin": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Admin access token (aud=admin)"
      }
    },
    "schemas": {
      "GoogleAuthDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshDto": {
        "type": "object",
        "properties": {}
      },
      "AdminLoginDto": {
        "type": "object",
        "properties": {}
      },
      "TotpVerifyDto": {
        "type": "object",
        "properties": {}
      },
...
      "AdjustBalanceDto": {
        "type": "object",
        "properties": {}
      },
      "BanUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAdminDto": {
        "type": "object",
        "properties": {}
      },
      "ResolveFraudDto": {
        "type": "object",
        "properties": {}
      },
      "RoundStartDto": {
        "type": "object",
        "properties": {}
      },
      "RoundCompleteDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}